Tie dewar GND to interface GND
It's clear that all grounds in the dewar (on the personality card and
the side cards) are all tied together.  So, there isn't any point, as
far as I know, to maintaining independent wiring on the interface card.
This commit removes independent grounds for each side and relabels
everything to the GND plane.