Update Shutter Control
Now I'm not so sure about the last commit.  This patch actually
completes the assignment of all shutter control pins.  I'm not sure it's
right, but it is complete, now.

I also changed FLASH_STB to a HV high current bias, as I think it is
actually the current source for the LEDs.  15V and 5V lines are still HV
high current also, eventhough I think only two high current supplies are
required for the shutter board.  I asked Denis about it.