Fix High Current Bias for SHUT+15V
Shutter +15V may need 475mA.  Only the Low Voltage high current biases
are capable of this (max 500mA); the HV high current biases are limited
to 250mA.  This patch fixes that, I think, by borrowing one of the LVHC
biases previously assigned to the pre-amp 5V, and assigning it to the
shutter +15V.  In exchange, the two pre-amp 5V lines are assigned HVHC
biases.  This line doesn't need to be as clean as other lines to the
pre-amp board, since the 5V only powers several relays.