Try Footprint Association in Schmatic
While most of the components on the schematic had footprints assigned,
they were only names to be matched later, I think, not direct links to
footprints in specific libraries.  When I went to CvPcb, it complained
bitterly about not being able to find the ccdcamera library, which I never
put into git.  Regardless, Windows and Linux will not agree on the
location of these footprints and futher adjustment is probably necessary.