Update Layout
Test of schematic change propagation succeeded.  That seems to work and I
understand how.

Placed DB37 connectors with reasonable spacing of .75 inch.  Placed spare
pin pads between.

Condensed the overall layout to try to keep the board width confined to
roughly the width of the Archon connectors, as a guide.  We probably want
to make the board slightly wider when we finalize.