Reroute After GND Plane, Fat Heater Traces
I used freerouter after creating the GND plane, and that just caused it
not to connect any of the GND pins.  That made the resulting traces better
(fewer vias), it didn't connect some pins at all.

Also, I figured out how to change the trace width for the heater circuit.