Delete Old DRV Traces
After updating the schematic and exporting the new netlist, I updated the
PCB.  I thought this would delete all the old traces, but I had to do some
of that myself.  In any case, I think this is ready for me to route the
new assignments for the driver modules.