Place Chassis Dimensions
The board was too big!  Argh.  I've started to redesign the whole
thing by recalcuating the board size.  To do this, I placed edges
for the Archon Chassis around the board located by the placement
holes of the ERM8 connectors.

Next, I placed new edge cuts, including the notch that I'll need
to work around for the new board design.