Finish Main Rerouting After Edge, High Current Fixes
This is a milestone.  After updating the size/edge of the board
and correcting the erroneous connections that were supposed to be
high current biases, I have *roughly* placed most of the traces.
I still need to finish the Heater traces and adjust all of the traces
for ideal spacing.

First, however, I'm going to expose GPIO pins; then I'll make the final
passes on the routes.