Fix OSHPark: Remove Extra Layer Alignment Marks
Aha!  I figured it out.  OSHPark couldn't detect the correct size of the
board because of the layer extra alignment marks I was using to mark the
dimensions of the panel space.  I moved the dimension measurements off
the margin layer, which I now think was not necessary, and deleted these
marks. Now the price is correct.